raster and tools services renamed. application and env params added when necessary
