+++ utils/widgets/WidgetHelper.js
@@ -13,7 +13,7 @@ import ScatterChart from 'utils/widgets/scatter';
 import getQueryByFilters from 'utils/getQueryByFilters';
 
 // Services
-import RasterService from 'services/RasterService';
+import RasterService from 'services/raster';
 
 const CHART_TYPES = {
   bar: BarChart,
